get metagame decks from goldfish

+++ getCustomDecksMTGgoldfish.ps1
@@ -1,17 +1,23 @@
 #download deck from mtggoldfish homepage pauper
 #
+$type="pauper"
 $site="https://www.mtggoldfish.com"
-$urlo="/deck/custom/pauper?page=10#paper"    # for more pages deck/custom/standard?page=2#paper
-$savepath="decks\"
+$urlo="/deck/custom/"+$type+"?page="    # add page number for more pages deck/custom/standard?page=2#paper
+$savepath="decks\goldfish\custom\"+$type+"\"
 
+New-item -type directory -path $savepath -Force
 
 $uripathDownload="/deck/download/"
 $regex = "/deck/([0-9]+)#paper"
 #https://www.mtggoldfish.com/deck/download/5267732  to download txt format
 
+for ($i=1;$i -le 20;$i++) {
+    $uri=$site+$urlo+$i+"#paper"
+    Write-Host $uri
+    $decklist+=(Invoke-WebRequest –Uri $uri ).Links|  Where-Object {$_.href -like “/deck/*#paper"}|Select-Object -ExpandProperty href
+}
+Write-Host $decklist
 
-$uri=$site+$urlo
-$decklist=(Invoke-WebRequest –Uri $uri ).Links|  Where-Object {$_.href -like “/deck/*#paper"}|Select-Object -ExpandProperty href
 #((Invoke-WebRequest –Uri $urlo).Links | Where-Object {$_.href -like “/deck"})
 #Write-Host $decklist
 
@@ -30,10 +36,15 @@ foreach ($num in $numbers) {
     $uri=$site+$uripathDownload+$num
     #Write-Host $uri
     #don't like this method to extract but it was the fastest
-    $head=[string](Invoke-WebRequest –Uri $uri).Headers['Content-Disposition']
+    $head=[string]((Invoke-WebRequest –Uri $uri).Headers['Content-Disposition'])
     $filename = $savepath+$num+"-"+$head.Substring($head.IndexOf("=")+1).Replace('"','').Replace('/','_')
     Write-Host $filename
-    (Invoke-WebRequest –Uri $uri).Content | Out-File -FilePath $filename
+    if ('' -eq ($head.Substring($head.IndexOf("=")+1).Replace('"','').Replace('/','_')) ) {
+        Write-Warning "no Content-Disposition Header for  $uri"
+    } else {
+        (Invoke-WebRequest –Uri $uri).Content | Out-File -FilePath $filename
+    }
+   
 
 }
